feat: fetch models after adding OpenAI key
